# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Healthcare Platform API
  version: 1.0.0
  description: |
    Comprehensive API for the Global Healthcare SaaS Platform providing telemedicine,
    patient management, appointments, clinical documentation, and billing services.

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain tokens via `/auth/login` or `/auth/register`.

    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per IP address.

  contact:
    name: API Support
    email: support@unifiedhealthcare.com
  license:
    name: Proprietary
    url: https://unifiedhealthcare.com/terms
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api-dev.unifiedhealthcare.com/api/v1
    description: Development environment
  - url: https://api-staging.unifiedhealthcare.com/api/v1
    description: Staging environment
  - url: https://api.unifiedhealthcare.com/api/v1
    description: Production environment
...
security:
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get authenticated user information
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /roles:
    get:
      tags:
        - Authentication
      summary: List available roles
      description: Get list of all system roles (admin only)
      operationId: getRoles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    name:
                      type: string
                      example: provider
                    description:
                      type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # ==========================================
  # User Endpoints
  # ==========================================
...
  /encounters:
    post:
      tags:
        - Encounters
      summary: Create clinical encounter
      description: Create a new clinical encounter (provider/admin only)
      operationId: createEncounter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEncounterRequest'
      responses:
        '201':
          description: Encounter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    get:
      tags:
        - Encounters
      summary: List encounters
      description: List encounters with optional filtering
      operationId: listEncounters
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by patient ID
        - name: providerId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by provider ID
      responses:
        '200':
          description: List of encounters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EncounterResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /encounters/{id}:
    get:
      tags:
        - Encounters
      summary: Get encounter by ID
      description: Retrieve specific encounter details
      operationId: getEncounter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Encounter found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Encounters
      summary: Update encounter
      description: Update encounter status (provider/admin only)
      operationId: updateEncounter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEncounterRequest'
      responses:
        '200':
          description: Encounter updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /encounters/{id}/notes:
    post:
      tags:
        - Encounters
      summary: Add clinical note
      description: Add a clinical note to an encounter (provider/admin only)
      operationId: addClinicalNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddClinicalNoteRequest'
      responses:
        '201':
          description: Note added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicalNoteResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    get:
      tags:
        - Encounters
      summary: Get clinical notes
      description: Retrieve all clinical notes for an encounter
      operationId: getClinicalNotes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of clinical notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClinicalNoteResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /encounters/{id}/start:
    post:
      tags:
        - Encounters
      summary: Start encounter
      description: Mark encounter as started (provider/admin only)
      operationId: startEncounter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Encounter started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /encounters/{id}/end:
    post:
      tags:
        - Encounters
      summary: End encounter
      description: Mark encounter as finished (provider/admin only)
      operationId: endEncounter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Encounter ended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==========================================
  # Document Endpoints
  # ==========================================
...
  /consents:
    post:
      tags:
        - Consent
      summary: Create consent record
      description: Record patient consent
      operationId: createConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsentRequest'
      responses:
        '201':
          description: Consent recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /consents/{id}:
    get:
      tags:
        - Consent
      summary: Get consent record
      description: Retrieve consent record by ID
      operationId: getConsent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Consent record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        tokenType:
          type: string
          example: Bearer
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            firstName:
              type: string
            lastName:
              type: string
            role:
              type: string

    # ==========================================
    # User Schemas
    # ==========================================
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [patient, provider, admin]
        status:
          type: string
          enum: [active, inactive, suspended]
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateEncounterRequest:
      type: object
      required:
        - patientId
        - providerId
        - type
      properties:
        patientId:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
        type:
          type: string
          enum: [virtual, in_person, phone]
        reasonForVisit:
          type: string
          minLength: 1
          maxLength: 1000
    UpdateEncounterRequest:
      type: object
      properties:
        status:
          type: string
          enum: [planned, in_progress, finished, cancelled]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
    AddClinicalNoteRequest:
      type: object
      required:
        - noteType
        - content
      properties:
        noteType:
          type: string
          enum: [progress, assessment, plan, chief_complaint, history, physical_exam, diagnosis, other]
        content:
          type: string
          minLength: 1
          maxLength: 50000
    EncounterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        appointmentId:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [virtual, in_person, phone]
        status:
          type: string
          enum: [planned, in_progress, finished, cancelled]
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: array
          items:
            $ref: '#/components/schemas/ClinicalNoteResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ClinicalNoteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        noteType:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time

    # ==========================================
    # Document Schemas
    # ==========================================
...
    CreateConsentRequest:
      type: object
      required:
        - patientId
        - type
        - granted
      properties:
        patientId:
          type: string
          format: uuid
        type:
          type: string
          enum: [data-sharing, treatment, marketing, research]
        granted:
          type: boolean
        scope:
          type: string
        expiresAt:
          type: string
          format: date-time
    ConsentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        type:
          type: string
        granted:
          type: boolean
        scope:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ==========================================
    # Common Schemas
    # ==========================================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time

  # ==========================================
  # Reusable Responses
  # ==========================================
  responses:
    BadRequestError:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Bad Request"
            message: "Invalid input data"
            timestamp: "2024-01-01T00:00:00Z"
    UnauthorizedError:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Authentication token is missing or invalid"
            timestamp: "2024-01-01T00:00:00Z"
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Forbidden"
            message: "You do not have permission to access this resource"
            timestamp: "2024-01-01T00:00:00Z"
    NotFoundError:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not Found"
            message: "The requested resource was not found"
            timestamp: "2024-01-01T00:00:00Z"
...